+++ src/Util/GlideUrl.php
@@ -4,6 +4,7 @@
 namespace AmpedWeb\GlideInABox\Util;
 
 
+use AmpedWeb\GlideInABox\Traits\Adjustments;
 use AmpedWeb\GlideInABox\Traits\Crop;
 use AmpedWeb\GlideInABox\Traits\Effects;
 use AmpedWeb\GlideInABox\Traits\Encode;
@@ -16,7 +17,7 @@ use League\Glide\Urls\UrlBuilderFactory;
 
 class GlideUrl
 {
-    use Orientation, Flip, Crop, Size, Effects, Encode;
+    use Orientation, Flip, Crop, Size, Adjustments, Effects, Encode;
 
     /**
      * The filepath of our image being manipulated